Apply suggested modifications
CHANGES:

- Change from `yyyy.mm.dd` version schema to SemVer;
- Move common functions into `utils.py`;
- Sort dependencies alphabetically in `package.xml`.
