+++ racecar_beacon/setup.py
@@ -12,7 +12,7 @@ executables = [
 
 setup(
     name=package_name,
-    version="2025.4.12",
+    version="1.0.0",
     packages=find_packages(exclude=["test"]),
     data_files=[
         ("share/ament_index/resource_index/packages", ["resource/" + package_name]),
